code standard updates after pylint
I moved part of the main function to the extract file. It was telling me we had too many branches in the function and they should be slit up. The rest of it was mostly just getting rid of white space or getting rid of ambiguous except statements. Other changes we just updating any variable like filter or all because those are built in functions in python and it was overwriting them. Will do the flake8 tomorrow.